Rename package to .core
