--- eventmesh-dashboard-core/src/main/java/org/apache/eventmesh/dashboard/core/Main.java
@@ -1,7 +1,0 @@
-package org.apache.eventmesh.dashboard.core;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
