+++ eventmesh-dashboard-core/src/main/java/org/apache/eventmesh/dashboard/core/exception/EventMeshAdminException.java
@@ -15,9 +15,9 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.dashboard.exception;
+package org.apache.eventmesh.dashboard.core.exception;
 
-import org.apache.eventmesh.dashboard.enums.Status;
+import org.apache.eventmesh.dashboard.core.enums.Status;
 
 /**
  * EventMeshAdmin Application side exception
