+++ eventmesh-dashboard-core/src/main/java/org/apache/eventmesh/dashboard/core/service/Main.java
@@ -1,4 +1,4 @@
-package org.apache.eventmesh.dashboard.service;
+package org.apache.eventmesh.dashboard.core.service;
 
 public class Main {
     public static void main(String[] args) {
